+++ src/styles/globals.scss
@@ -31,7 +31,7 @@ html {
 body {
   border: 1px solid black;
   width: 100%;
-  height: 40rem;
+  height: 100%;
   display: grid;
   grid-template-rows: 1fr 12fr;
   grid-template-columns: 1fr;
