--- src/styles/signup.scss
@@ -1,44 +1,0 @@
-@import './globals.scss';
-
-.signup-form {
-  width: 80%;
-  margin: 0 auto;
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-  display: flex;
-  flex-direction: column;
-}
-
-.signup-form__label {
-  margin-bottom: 0.5rem;
-}
-
-.signup-form__input {
-  height: 1.7rem;
-  margin-bottom: 1.2rem;
-  padding-left: 0.4rem;
-  border: 1px solid #000;
-  border-radius: 5px;
-  font-size: 1rem;
-}
-
-.signup-message {
-  margin-bottom: 0.3rem;
-  color: #ff0000;
-  font-size: 0.8rem;
-}
-
-.submit-button {
-  width: 100%;
-  padding: 0.4rem;
-  font-size: 1.3rem;
-  font-weight: 500;
-  background-color: $main-color;
-  color: #fff;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.submit-button:disabled {
-  background-color: #ddd;
-}
