+++ src/styles/templates/category/category.scss
