+++ src/styles/templates/home.scss
