+++ src/styles/templates/search/search.scss
@@ -1,4 +1,4 @@
-@import '@/styles/mixin';
+@import '@/styles/search/mixin';
 
 #searchPage {
   .search {
