+++ src/templates/login/loginPage.tsx
@@ -1,5 +1,5 @@
 import Btn from '@/components/btn';
-import '@/styles/app/login.scss';
+import '@/styles/templates/login/login.scss';
 import Link from 'next/link';
 
 export default function LoginPage() {
@@ -13,7 +13,7 @@ export default function LoginPage() {
           <p>비밀번호(PW)</p>
           <input />
         </div>
-        <Btn href="/main" label="로그인" />
+        <Btn href="/main" label="로그인" disabled={false} />
         <div className="login__pw-text">
           <p>비밀번호를 잊어버리셨나요?</p>
           <Link className="pw" href="login">
